Rename building panel modals
